pyproject.toml
[project]
name = "hud-spreadsheetbench"
version = "0.1.0"
description = "HUD SpreadsheetBench MCP Server"
authors = [{name = "HUD Team"}]
requires-python = ">=3.11"
dependencies = [
    "hud-python",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "xlsxwriter>=3.1.0",
    "jupyter-client>=8.0.0",
    "jupyter-kernel-gateway>=3.0.0",
    "ipython>=8.0.0",
    "nbformat>=5.7.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "tornado>=6.0.0",
    "aiohttp>=3.8.0",
]